🚚 Rename APS 응용 Day08 - 10
